Move middlewares to middlewares.ts
