+++ examples/node-http-server-ts/src/middlewares.ts
@@ -1,11 +1,40 @@
 import * as qs from 'qs';
 import { ServerResponse, IncomingMessage } from 'http';
+
 import * as store from '@leapfrogtechnology/async-store';
+import AsyncStoreAdapter from '@leapfrogtechnology/async-store/dist/AsyncStoreAdapter';
 
 import * as logger from './logger';
 import { doSomething } from './service';
 
 export type Middleware = (req: IncomingMessage, res: ServerResponse) => void;
+
+/**
+ * Handle incoming http request.
+ *
+ * @param {IncomingMessage} req
+ * @param {ServerResponse} res
+ */
+function handleRequest(req: IncomingMessage, res: ServerResponse) {
+  const params = (req.url || '').split('?', 2)[1];
+
+  store.set({ query: params });
+
+  requestParams(req, res);
+  add(req, res);
+}
+
+/**
+ * Listener function for the incoming http requests.
+ *
+ * @param {IncomingMessage} req
+ * @param {ServerResponse} res
+ */
+export function initializeApp(req: IncomingMessage, res: ServerResponse) {
+  const init = store.initialize(AsyncStoreAdapter.DOMAIN);
+
+  init(() => handleRequest(req, res), { req, res, error: logger.error });
+}
 
 /**
  * Middleware to set query params `a` and `b` on async-store.
