+++ examples/node-http-server-ts/src/service.ts
@@ -2,8 +2,6 @@ import * as logger from './logger';
 
 /**
  * An example function making use of the request context set in the store.
- *
- * @returns {void}
  */
 export function doSomething() {
   // Do something with the request.
